project/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
python-multipart==0.0.6
pillow==10.1.0
requests==2.31.0 
psutil==5.9.8
py-notifier==0.0.1
win10toast==0.9
fpdf==1.7.2
pdf2image==1.16.3
python-docx==0.8.11
poppler-utils==0.1.0
